AzureAppMigration Submission/techconfQueueTrigger/requirements.txt
azure-functions==1.7.2
psycopg2-binary==2.9.1
python-http-client==3.3.3
sendgrid==6.8.2
starkbank-ecdsa==1.1.1
